Improvements to styling of VolumeSlider & add interaction
